+++ src/backend/joanie/tests/core/api/admin/orders/test_retrieve.py
@@ -202,13 +202,7 @@ class OrdersAdminApiRetrieveTestCase(TestCase):
                     "type": order.main_invoice.type,
                     "updated_on": format_date(order.main_invoice.updated_on),
                 },
-                "credit_card": {
-                    "id": str(order.credit_card.id),
-                    "last_numbers": order.credit_card.last_numbers,
-                    "brand": order.credit_card.brand,
-                    "expiration_month": order.credit_card.expiration_month,
-                    "expiration_year": order.credit_card.expiration_year,
-                },
+                "credit_card": None,
                 "has_waived_withdrawal_right": order.has_waived_withdrawal_right,
             },
         )